+++ POS/POS/Modules/Reports/ViewModels/ReportViewModel.cs
@@ -1,9 +1,12 @@
 ﻿using BoldReports.UI.Xaml;
+using BoldReports.Windows;
 using Framework.Core.FileSystem;
+using Framework.Infrastructure.Identity.Data;
 using Framework.WPF.Modules.CaliburnMicro;
 using Framework.WPF.ScreenManagement;
 using POS.Core.Reports;
 using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using System.Windows;
 
@@ -11,13 +14,22 @@ namespace POS.Modules.Reports.ViewModels
 {
     public class ReportViewModel : ExtendedScreenBase
     {
+        private const string REPORT_DATASOURCE_NAME = "LotteryRetail";
+        private const string REPORT_PATH = @"Resources\Reports\{0}.rdl";
+
         private readonly IFilePathService _filePathService;
         private readonly ReportContext _reportContext;
+        private readonly ISecurityDbConnectionInfo _securityDbConnectionInfo;
 
-        public ReportViewModel(IScreenServices screenManagementServices, IFilePathService filePathService , ReportContext reportContext) : base(screenManagementServices)
+        public ReportViewModel(
+            ReportContext reportContext,
+            IScreenServices screenManagementServices, 
+            IFilePathService filePathService,             
+            ISecurityDbConnectionInfo securityDbConnectionInfo) : base(screenManagementServices)
         {
-            _filePathService = filePathService;
             _reportContext = reportContext;
+            _filePathService = filePathService;            
+            _securityDbConnectionInfo = securityDbConnectionInfo;
         }
 
         public async Task HandleLoaded(RoutedEventArgs eventArgs)
@@ -25,8 +37,18 @@ namespace POS.Modules.Reports.ViewModels
             if (!(eventArgs.OriginalSource is ReportViewer reportViewer)) return;
             try
             {
-                reportViewer.ReportPath = _filePathService.Combine(Environment.CurrentDirectory, $@"Resources\Reports\{_reportContext.SelectedReportName.Name}.rdl");
+                reportViewer.ReportPath = _filePathService.Combine(Environment.CurrentDirectory, String.Format(REPORT_PATH, _reportContext.SelectedReportName.Name));
                 reportViewer.RefreshReport();
+
+                reportViewer.SetDataSourceCredentials(new List<DataSourceCredentials> {
+                    new DataSourceCredentials
+                        {
+                            ConnectionString = _securityDbConnectionInfo.GetConnectionString(),
+                            Name = REPORT_DATASOURCE_NAME,
+                            IntegratedSecurity = true
+                        }
+                });               
+
             }
             catch (Exception exception)
             {